refactoring each auth method to different partial class files in subfolder
